create component

+++ src/components/SignUp/index.js
@@ -1,16 +1,35 @@
 import React, { useState } from "react";
+import { Button } from "react-bootstrap";
+import "./sign.css";
 const SignUp = () => {
   const [name, setName] = useState("");
   const [age, setAge] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [passwordConfirmation, setPasswordConfirmation] = useState("");
+  const [inValid, setInValid] = useState("");
+
+  const validateEmail = (email) => {
+    console.log("llllllllllllll");
+    let regexEmail = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
+    if (email.match(regexEmail)) {
+      return email;
+    } else {
+      setInValid("is-invalid");
+    }
+  };
+
   const register = () => {
-    localStorage.setItem("name", name);
-    localStorage.setItem("age", age);
-    localStorage.setItem("email", email);
-    localStorage.setItem("password", password);
-    localStorage.setItem("passwordConfirmation", passwordConfirmation);
+    console.log("llllllllllllllllllllllllllllllllllllllllllllll");
+    if (name && age && email && password && passwordConfirmation) {
+      localStorage.setItem("name", name);
+      localStorage.setItem("age", age);
+      localStorage.setItem("email", email);
+      localStorage.setItem("password", password);
+      localStorage.setItem("passwordConfirmation", passwordConfirmation);
+    } else {
+      setInValid("Fill add fields");
+    }
   };
   return (
     <div>
@@ -40,6 +59,7 @@ const SignUp = () => {
           <input
             type="email"
             placeholder="Enter email"
+            onChange={validateEmail}
             onChange={(e) => {
               setEmail(e.target.value);
             }}
@@ -65,10 +85,13 @@ const SignUp = () => {
             }}
           />
         </div>
-        <button variant="primary" type="submit" onClick={register}>
+        <Button variant="primary" onClick={register}>
           Sign Up
-        </button>
+        </Button>
       </form>
+      {!name || !age || !email || !password || !passwordConfirmation ? (
+        <div className="fill">{inValid}</div>
+      ) : null}
     </div>
   );
 };
